Move more things into template-ui
