+++ src/core/settings.ts
@@ -1,6 +1,6 @@
 import { Template } from '../types/types';
 import { loadTemplates, createDefaultTemplate, templates, getTemplates, findTemplateById, saveTemplateSettings } from '../managers/template-manager';
-import { updateTemplateList, showTemplateEditor } from '../managers/template-ui';
+import { updateTemplateList, showTemplateEditor, resetUnsavedChanges } from '../managers/template-ui';
 import { initializeGeneralSettings, addVault } from '../managers/general-settings';
 import { initializeSidebar } from '../utils/ui-utils';
 import { initializeDragAndDrop } from '../utils/drag-and-drop';
@@ -8,7 +8,6 @@ import { initializeAutoSave } from '../utils/auto-save';
 import { exportTemplate, importTemplate, initializeDropZone } from '../utils/import-export';
 import { createIcons } from 'lucide';
 import { icons } from '../icons/icons';
-import { resetUnsavedChanges } from '../managers/template-manager';
 
 function updateUrl(section: string, templateId?: string): void {
 	let url = `${window.location.pathname}?section=${section}`;
